pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chunck-norris-ai"
version = "0.1"
dependencies = [
    "Pillow>=11.2.1",
    "openai>=1.75.0",
    "falkordb>=1.1.1",
    "python-docx>=1.1.2",
    "mammoth>=1.9.0",
    "redis>=5.0.1",
] 